[MM, JW] Committing progress

+++ app/src/main/java/rhyme/a/is/nine/foodmanager/product/PriceEntity.java
@@ -8,10 +8,10 @@ import java.util.Date;
 public class PriceEntity {
 
     private String name = null;
-    private double price = -1;
+    private float price = -1;
     private Date buyDate = null;
 
-    public PriceEntity(String name, double price, Date buyDate) {
+    public PriceEntity(String name, float price, Date buyDate) {
         this.name = name;
         this.price = price;
         this.buyDate = buyDate;
@@ -25,11 +25,11 @@ public class PriceEntity {
         this.name = name;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
